+++ app/Http/Controllers/Api/PostController.php
@@ -23,8 +23,12 @@ class PostController extends Controller
 
     public function store(PostStoreRequest $request): PostResource
     {
-        $post = $this->service->create($request->validated());
+        $data = $request->validated();
+        $data['user_id'] = $request->user()->id;
+
+        $post = $this->service->create($data);
         $post->loadCount('comments')->load('user');
+
         return new PostResource($post);
     }
 
@@ -36,13 +40,18 @@ class PostController extends Controller
 
     public function update(PostUpdateRequest $request, Post $post): PostResource
     {
+        $this->authorize('update', $post);
+
         $post = $this->service->update($post, $request->validated());
         $post->loadCount('comments')->load('user');
+
         return new PostResource($post);
     }
 
     public function destroy(Post $post)
     {
+        $this->authorize('delete', $post);
+
         $this->service->delete($post);
         return response()->noContent();
     }
